# generated/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/identity-access-reviews": {
      "get": {
        "operationId": "AdminIdentityAccessController_listReviews",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Identity access reviews visible to the current admin scope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "identityAccessReviews"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "identityAccessReviews": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "userId",
                          "displayName",
                          "email",
                          "provider",
                          "providerSubject",
                          "status",
                          "scopeOrganizationUnitId",
                          "scopeOrganizationUnitName",
                          "requestedRole",
                          "assignedRole",
                          "expiresAt",
                          "decidedBy",
                          "decidedAt",
                          "decisionNote",
                          "createdAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false,
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "userId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string",
                            "format": "email"
                          },
                          "provider": {
                            "type": "string"
                          },
                          "providerSubject": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "pending",
                              "confirmed",
                              "rejected",
                              "expired"
                            ]
                          },
                          "scopeOrganizationUnitId": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "scopeOrganizationUnitName": {
                            "type": "string",
                            "nullable": true
                          },
                          "requestedRole": {
                            "type": "string",
                            "enum": [
                              "CANDIDATE",
                              "BROTHER",
                              "OFFICER",
                              "SUPER_ADMIN"
                            ],
                            "nullable": true
                          },
                          "assignedRole": {
                            "type": "string",
                            "enum": [
                              "CANDIDATE",
                              "BROTHER",
                              "OFFICER",
                              "SUPER_ADMIN"
                            ],
                            "nullable": true
                          },
                          "expiresAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "decidedBy": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          },
                          "decidedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "decisionNote": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "The current user does not have Admin Lite access.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "details",
                        "requestId",
                        "timestamp"
                      ],
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {}
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-identity-access"
        ]
      }
    },
    "/api/admin/identity-access-reviews/{id}": {
      "get": {
        "operationId": "AdminIdentityAccessController_getReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Identity access review detail visible to the current admin scope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "identityAccessReview"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "identityAccessReview": {
                      "type": "object",
                      "required": [
                        "id",
                        "userId",
                        "displayName",
                        "email",
                        "provider",
                        "providerSubject",
                        "status",
                        "scopeOrganizationUnitId",
                        "scopeOrganizationUnitName",
                        "requestedRole",
                        "assignedRole",
                        "expiresAt",
                        "decidedBy",
                        "decidedAt",
                        "decisionNote",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "userId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "provider": {
                          "type": "string"
                        },
                        "providerSubject": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "pending",
                            "confirmed",
                            "rejected",
                            "expired"
                          ]
                        },
                        "scopeOrganizationUnitId": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        },
                        "scopeOrganizationUnitName": {
                          "type": "string",
                          "nullable": true
                        },
                        "requestedRole": {
                          "type": "string",
                          "enum": [
                            "CANDIDATE",
                            "BROTHER",
                            "OFFICER",
                            "SUPER_ADMIN"
                          ],
                          "nullable": true
                        },
                        "assignedRole": {
                          "type": "string",
                          "enum": [
                            "CANDIDATE",
                            "BROTHER",
                            "OFFICER",
                            "SUPER_ADMIN"
                          ],
                          "nullable": true
                        },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "decidedBy": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        },
                        "decidedAt": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        },
                        "decisionNote": {
                          "type": "string",
                          "nullable": true
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "The current user does not have permission to view this review.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "details",
                        "requestId",
                        "timestamp"
                      ],
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {}
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "The identity access review is not visible in the current scope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "details",
                        "requestId",
                        "timestamp"
                      ],
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {}
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-identity-access"
        ]
      }
    },
...
    "/api/admin/identity-access-reviews/expire": {
      "post": {
        "operationId": "AdminIdentityAccessController_expirePendingReviews",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Expired pending identity access reviews past their 30-day window.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "expired"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "expired": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Only Super Admin can run global identity review expiry.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "details",
                        "requestId",
                        "timestamp"
                      ],
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {}
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-identity-access"
        ]
      }
    },
...
  },
  "info": {
    "title": "JP2 Knights API",
    "description": "V1 REST API for the JP2 Knights app ecosystem.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}